+++ facia/app/controllers/FrontController.scala
@@ -5,7 +5,6 @@ import front._
 import model._
 import conf._
 import play.api.mvc._
-import model.Trailblock
 import scala.Some
 
 import concurrent.Future
