--- facia/app/facia/editions/Us.scala
@@ -1,55 +1,0 @@
-package facia.editions
-
-import facia.Edition
-import common._
-import org.joda.time.DateTimeZone
-import model.MetaData
-import common.NavItem
-import contentapi.QueryDefaults
-
-object Us extends Edition(
-  id = "US",
-  displayName = "US edition",
-  timezone = DateTimeZone.forID("America/New_York"),
-  hreflang = "en-us") with Sections with Zones with QueryDefaults {
-
-  implicit val US = Us
-  val zones = Seq(
-    newsZone,
-    sportsZone,
-    cifZone,
-    cultureZone,
-    technologyZone,
-    businessZone,
-    moneyZone,
-    lifeandstyleZone,
-    travelZone
-  )
-
-  def navigation(metadata: MetaData) = Seq(
-    NavItem(home),
-    NavItem(us),
-    NavItem(world, Seq(us, europe, middleeast, asiapacific, africa, americas)),
-    NavItem(sports, Seq(nfl, mlb, nba, mls, nhl, football)),
-    footballNav(metadata),
-    NavItem(cif, Seq(cifbelief, cifgreen)),
-    NavItem(lifeandstyle, Seq(foodanddrink, fashion, relationships, healthandwellbeing, women)),
-    NavItem(culture, Seq(film, music, books, televisionandradio, artanddesign, stage)),
-    NavItem(business, Seq(economics, banking, property, workandcareers, savings)),
-    NavItem(technology, Seq(internet, games, mobilephones, appsblog)),
-    NavItem(environment, Seq(climatechange, wildlife, energy, conservation, food)),
-    NavItem(media)
-  )
-
-
-  val configuredFronts = Map(
-    Editionalise("", Us) -> Seq(),
-
-    Editionalise("sport", Us) -> Seq(),
-
-    Editionalise("culture", Us) -> Seq(),
-
-    Editionalise("australia", Us)  -> Seq()
-  )
-
-}
